Add Fronts banner ads to Tag Fronts.
